pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "recbar"
version = "1.2.0"
description = "Lightweight OBS recording companion bar — indicator, mic VU, chapters, auto-scene, mobile remote"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "rathin"},
]
keywords = ["obs", "recording", "streaming", "indicator", "bar", "obs-studio"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "PyQt6>=6.5",
    "websocket-client>=1.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.3",
]

[project.urls]
Homepage = "https://github.com/KHET-1/recbar"
Issues = "https://github.com/KHET-1/recbar/issues"

[project.scripts]
recbar = "recbar.__main__:main"
recbar-ctl = "recbar.ctl:main"
recbar-test = "recbar.test_suite:main"

[tool.setuptools.packages.find]
include = ["recbar*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
